requirements.txt
python-kasa
HAP-python[QRCode]
pyyaml
